Anuja Assignment 57 in Anujas folder and delete from Amrutas folder
